Change components and containers to define (presentational) component props in it
